Add auth

+++ app/Http/Controllers/MeetingController.php
@@ -5,12 +5,17 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Meeting;
 use Carbon\Carbon;
+use JWTAuth;
 
 class MeetingController extends Controller
 {
     public function __construct()
     {
-
+        $this->middleware('jwt.auth', ['only' => [
+            'update',
+            'store',
+            'destroy'
+        ]]);
     }
 
     /**
@@ -49,13 +54,16 @@ class MeetingController extends Controller
             'title' => 'required',
             'description' => 'required',
             'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
         ]);
+
+        if (!$user = JWTAuth::parseToken()->authenticate()) {
+            return response()->json(['msg' => 'User not found', 404]);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = new Meeting([
             'time' => Carbon::createFromFormat('YmdHie', $time),
@@ -122,23 +130,16 @@ class MeetingController extends Controller
             'title' => 'required',
             'description' => 'required',
             'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
         ]);
+
+        if (!$user = JWTAuth::parseToken()->authenticate()) {
+            return response()->json(['msg' => 'User not found', 404]);
+        }
 
         $title =$request->input('title');
         $description =$request->input('description');
         $time =$request->input('time');
-        $user_id =$request->input('user_id');
-        $meeting = [
-            'title' => $title,
-            'description' => $description,
-            'time' => $time,
-            'user_id' => $user_id,
-            'view_meeting' =>[
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
-            ]
-        ];
+        $user_id = $user->id;
 
         $meeting = Meeting::with('users')->findOrFail($id);
 
@@ -180,6 +181,17 @@ class MeetingController extends Controller
     {
 
         $meeting = Meeting::findOrFail($id);
+
+        if (!$user = JWTAuth::parseToken()->authenticate()) {
+            return response()->json(['msg' => 'User not found', 404]);
+        }
+
+        if (!$meeting->users()->where('user_id', $user)->first()) {
+            return response()->json([
+                'msg' =>'user not registered for meeting, update not successful',
+            ], 401);
+        }
+
         $users = $meeting->users;
         $meeting->users()->detach();
         if (!$meeting->delete()) {
